23. 인스타그램 클론코딩 - 이미지 데이터 가져오기 (양방향매핑 이해)
1. UserService 에서 회원프로필을 불러올 때 먼저 유저의 데이터를 불러온다.
2. 유저의 데이터만으로는 이미지를 가져올 수 없기 때문에 양방향 매핑을 해준다.
3. private List<Image> images; 하면 images 컬럼이 만들어지는데 그럼 어떤 형인지 알 수 없다.
4. 그래서 mappedBy 를 통해 연관관계 주인을 Image로 넘기고 테이블 컬럼을 만들지 말라고 한다.
5. Lazy, Eager 에 따라서 image 불러오는 방식 차이가 있음

+++ src/main/java/com/cos/photogramstart/domain/user/UserRepository.java
@@ -6,9 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Map;
 
 //@Repository 어노테이션 없어도 IoC에 자동으로 등록 ( 상속받으면)
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     // JPA Query Method
     User findByUsername(String username);
-
-
 }
